requirements.txt
# ============================================================================
# Physics VLM SFT & RLHF Pipeline - Dependencies
# 基于 RTX 4090 (24GB) 训练环境
# ============================================================================

# -------------------- Core Deep Learning Frameworks --------------------
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers==5.4.0
accelerate==1.13.0

# -------------------- Model Training & Fine-tuning --------------------
# Parameter Efficient Fine-Tuning
peft==0.14.0

# Reinforcement Learning
trl==0.24.0

# Quantization & Memory Optimization
bitsandbytes==0.49.2

# -------------------- Attention Optimization --------------------
# Flash Attention 2 (需要编译安装，约15分钟)
flash-attn==2.8.3
xformers==0.0.28.post3

# -------------------- SWIFT Framework (RLHF Training) --------------------
ms-swift==4.0.2

# -------------------- Inference Engine --------------------
vllm==0.7.2
triton==3.1.0

# -------------------- Data Processing --------------------
datasets==3.6.0
pandas==3.0.1
pyarrow==23.0.1
numpy==1.26.4
Pillow==12.1.1

# -------------------- Qwen VLM Utils --------------------
qwen-vl-utils==0.0.14
opencv-python-headless==4.11.0.86
einops==0.8.2

# -------------------- Hugging Face Ecosystem --------------------
huggingface-hub==1.8.0
tokenizers==0.22.2
safetensors==0.7.0
sentencepiece==0.2.1

# -------------------- Training Monitoring --------------------
tensorboard==2.16.2
wandb==0.25.1
tqdm==4.67.3

# -------------------- Evaluation --------------------
evalscope==1.5.1
rouge-chinese==1.0.3
nltk==3.9.4
jieba==0.42.1

# -------------------- Utilities --------------------
PyYAML==6.0.3
fire==0.7.1
gradio==5.50.0
fastapi==0.135.2
uvicorn==0.42.0

# -------------------- Optional: LLaMA-Factory (SFT Training) --------------------
# 建议从源码安装:
# git clone https://github.com/hiyouga/LLaMA-Factory.git
# cd LLaMA-Factory && pip install -e .

# -------------------- Build Tools (for flash-attn) --------------------
# 编译 flash-attn 需要:
# - CUDA Toolkit >= 11.8
# - ninja (build tool)
# ninja==1.11.1  # 已在系统中
